# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multibank aggregation service for VTB Hackathon. Use /auth/register or /auth/login to get JWT token, then click \"Authorize\" button and paste token.",
        "title": "VTB Multibank Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@vtbmultibank.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/banks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all bank connections for authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Get user's bank connections",
                "responses": {
                    "200": {
                        "description": "List of connections",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/connect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create connection to a bank using bank client ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Connect bank account",
                "parameters": [
                    {
                        "description": "Bank connection details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ConnectBankRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Bank connected successfully",
                        "schema": {
                            "$ref": "#/definitions/bank.ConnectBankResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/{provider}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove bank connection for authenticated user",
                "tags": [
                    "banks"
                ],
                "summary": "Disconnect bank",
                "parameters": [
                    {
                        "enum": [
                            "vbank",
                            "sbank",
                            "abank"
                        ],
                        "type": "string",
                        "description": "Bank provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bank disconnected successfully"
                    },
                    "400": {
                        "description": "Invalid bank provider",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/{provider}/accounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all accounts from specified bank provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Get bank accounts",
                "parameters": [
                    {
                        "enum": [
                            "vbank",
                            "sbank",
                            "abank"
                        ],
                        "type": "string",
                        "description": "Bank provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of accounts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid bank provider",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/{provider}/accounts/{accountId}/balances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve balances for a specific account from a connected bank",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Get account balances",
                "parameters": [
                    {
                        "enum": [
                            "vbank",
                            "sbank",
                            "abank"
                        ],
                        "type": "string",
                        "description": "Bank provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "account_id": {
                                    "type": "string"
                                },
                                "balances": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.Balance"
                                    }
                                },
                                "bank_provider": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banks/{provider}/accounts/{accountId}/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve transaction history for a specific account from a connected bank",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Get account transactions",
                "parameters": [
                    {
                        "enum": [
                            "vbank",
                            "sbank",
                            "abank"
                        ],
                        "type": "string",
                        "description": "Bank provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "account_id": {
                                    "type": "string"
                                },
                                "bank_provider": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "transactions": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bank.ConnectBankResponse": {
            "type": "object",
            "properties": {
                "connection": {
                    "$ref": "#/definitions/models.BankConnection"
                },
                "consent_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.ConnectBankRequest": {
            "type": "object",
            "required": [
                "bank_client_id",
                "bank_provider"
            ],
            "properties": {
                "bank_client_id": {
                    "type": "string"
                },
                "bank_provider": {
                    "type": "string",
                    "enum": [
                        "vbank",
                        "sbank",
                        "abank"
                    ]
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Balance": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.BankConnection": {
            "type": "object",
            "properties": {
                "bank_client_id": {
                    "description": "team-xxx-x",
                    "type": "string"
                },
                "bank_provider": {
                    "$ref": "#/definitions/models.BankProvider"
                },
                "consent_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ConnectionStatus"
                },
                "token_expires_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.BankProvider": {
            "type": "string",
            "enum": [
                "vbank",
                "sbank",
                "abank"
            ],
            "x-enum-varnames": [
                "VBankProvider",
                "SBankProvider",
                "ABankProvider"
            ]
        },
        "models.ConnectionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "active",
                "expired",
                "revoked"
            ],
            "x-enum-varnames": [
                "ConnectionPending",
                "ConnectionActive",
                "ConnectionExpired",
                "ConnectionRevoked"
            ]
        },
...
        "models.Transaction": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "booking_date_time": {
                    "type": "string"
                },
                "credit_debit_indicator": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_information": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}